requirements.txt
click
jinja2
markdown
pyyaml
watchdog 